Refactor user creation and login into subdirectories and modules
