+++ app.js
@@ -5,8 +5,8 @@ var cookieParser = require('cookie-parser');
 var logger = require('morgan');
 
 var indexRouter = require('./routes/index');
-var usersRouter = require('./routes/api/v1/users');
-var sessionsRouter = require('./routes/api/v1/sessions');
+var usersRouter = require('./routes/api/v1/users/users');
+var sessionsRouter = require('./routes/api/v1/login/sessions');
 var forecastRouter = require('./routes/api/v1/forecast');
 
 var app = express();
