+++ routes/api/v1/login/sessions.js
@@ -1,10 +1,9 @@
 var express = require('express');
 var router = express.Router();
-var User = require('../../../models').User;
 const bcrypt = require('bcrypt');
-var srs = require('secure-random-string');
-var checkBody = require("./account_creation").checkBody;
-var findOrCreateUser = require("./account_creation").findOrCreateUser;
+var User = require('../../../../models').User;
+var checkBody = require("./login").checkBody;
+var tryLogin = require("./login").tryLogin;
 
 router.post('/', function(req, res, next) {
   res.setHeader("Content-Type", "application/json");
@@ -12,7 +11,7 @@ router.post('/', function(req, res, next) {
   if (payload) {
     res.status(400).send(JSON.stringify(payload));
   } else {
-    findOrCreateUser(req.body, function(response) {
+    tryLogin(req.body, function(response) {
       res.status(response[0]).send(JSON.stringify(response[1]));
     });
   }
